Bug 861039 - Update ANGLE to r2042. - r=bjacob
UltraBlame original commit: fe104053f4f21800b393ffde868d77ed8a2a3a3f

+++ gfx/angle/src/compiler/MapLongVariableNames.h
@@ -31,7 +31,7 @@ public:
     void Insert(const char* originalName, const char* mappedName);
 
     
-    int Size() const;
+    size_t Size() const;
 
 private:
     LongNameMap();
